Shared approval-batch detail; expandable approval rows on Notifications
Extract the approvals screen's batch rendering into
components/approval-batch.tsx — ApprovalBatchBody (requests / originating
script with CodeBlock / policy, owning the members+script sub-toggles and
the one-shot script fetch) and ThreadContextLine (#2372's status-at-the-
time line) — so the approvals queue/history and the Notifications view
render one implementation that cannot drift. The in-body thread link now
reads "Open thread" (was "Show thread") on both surfaces.

Notifications rows that carry a batch identity toggle an inline expansion
(collapsed by default, ActivityCard's useState toggle precedent, chevron
affordance): decision badge / "Awaiting decision", rejection reason,
thread-context line, and the full shared body; the chat deep-link moved
inside the expansion (the body's "Open thread", or an "Open in Approvals"
link for non-thread batches). Non-approval rows keep tap-to-navigate. The
batch fetch is a one-shot query by offset against the root stream, cached
forever once decided AND settled (deriveBatchDetail's `complete`), else on
the same 5s provisional refetch the thread-context line uses.
